pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genoexpect"
version = "0.1.0"
description = "Validate observed genomics metrics against expected study properties"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "OpenAI" }
]
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
genoexpect = "genoexpect.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
